[core] Follow-up to address comments of BaseException PR #55602
