+++ python/ray/tests/BUILD
@@ -510,6 +510,7 @@ py_test_module_list(
         "test_async_compat.py",
         "test_asyncio_cluster.py",
         "test_autoscaling_policy.py",
+        "test_baseexceptionandgroup.py",
         "test_bounded_unix_sockets.py",
         "test_component_failures.py",
         "test_concurrency_group.py",
@@ -519,7 +520,6 @@ py_test_module_list(
         "test_distributed_sort.py",
         "test_environ.py",
         "test_error_ray_not_initialized.py",
-        "test_exceptiongroup.py",
         "test_gcs_pubsub.py",
         "test_get_or_create_actor.py",
         "test_grpc_client_credentials.py",
