+++ python/ray/tests/test_baseexceptionandgroup.py
@@ -1,18 +1,132 @@
+import pytest
 import sys
 from textwrap import dedent
 
-import pytest
-
 import ray
-from ray.exceptions import RayTaskError
+from ray.exceptions import (
+    RayTaskError,
+    ActorDiedError,
+    TaskCancelledError,
+    WorkerCrashedError,
+)
 
-pytestmark = pytest.mark.skipif(
+
+def test_baseexception_task(ray_start_regular_shared):
+    class MyBaseException(BaseException):
+        pass
+
+    @ray.remote
+    def task():
+        raise MyBaseException("abc")
+
+    with pytest.raises(MyBaseException):
+        ray.get(task.remote())
+
+
+def test_baseexception_actor_task(ray_start_regular_shared):
+    class MyBaseException(BaseException):
+        pass
+
+    @ray.remote
+    class Actor:
+        def f(self):
+            raise MyBaseException("abc")
+
+        async def async_f(self):
+            raise MyBaseException("abc")
+
+    a = Actor.remote()
+    with pytest.raises(MyBaseException):
+        ray.get(a.f.remote())
+
+    a = Actor.remote()
+    with pytest.raises(MyBaseException):
+        ray.get(a.async_f.remote())
+
+
+def test_baseexception_actor_creation(ray_start_regular_shared):
+    class MyBaseException(BaseException):
+        pass
+
+    @ray.remote
+    class Actor:
+        def __init__(self):
+            raise MyBaseException("abc")
+
+    try:
+        a = Actor.remote()
+        ray.get(a.__ray_ready__.remote())
+        raise Exception("abc")
+    except ActorDiedError as e:
+        assert "MyBaseException" in str(e)
+
+
+def test_baseexception_streaming_generator(ray_start_regular_shared):
+    class MyBaseException(BaseException):
+        pass
+
+    @ray.remote
+    def raise_at_beginning():
+        raise MyBaseException("rip")
+        yield 1
+
+    raise_at_beginning_ref = raise_at_beginning.remote()
+    with pytest.raises(MyBaseException):
+        ray.get(next(raise_at_beginning_ref))
+
+    @ray.remote
+    def raise_at_middle():
+        for i in range(1, 10):
+            if i == 5:
+                raise MyBaseException("rip")
+            yield i
+
+    raise_at_middle_ref = raise_at_middle.remote()
+    for i in range(1, 5):
+        assert i == ray.get(next(raise_at_middle_ref))
+    with pytest.raises(MyBaseException):
+        ray.get(next(raise_at_middle_ref))
+
+    @ray.remote(_generator_backpressure_num_objects=1)
+    def raise_after_backpressure():
+        for i in range(1, 10):
+            if i == 5:
+                raise MyBaseException("rip")
+            yield i
+
+    raise_after_backpressure_ref = raise_after_backpressure.remote()
+    for i in range(1, 5):
+        assert i == ray.get(next(raise_after_backpressure_ref))
+    with pytest.raises(MyBaseException):
+        ray.get(next(raise_after_backpressure_ref))
+
+
+def test_raise_system_exit(ray_start_regular_shared):
+    @ray.remote
+    def task():
+        raise SystemExit("abc")
+
+    with pytest.raises(WorkerCrashedError):
+        ray.get(task.remote())
+
+
+def test_raise_keyboard_interrupt(ray_start_regular_shared):
+    @ray.remote
+    def task():
+        raise KeyboardInterrupt("abc")
+
+    with pytest.raises(TaskCancelledError):
+        ray.get(task.remote())
+
+
+skip_if_python_less_than_3_11 = pytest.mark.skipif(
     sys.version_info < (3, 11),
     reason="ExceptionGroup is only available in Python 3.11+",
 )
 
 
-def test_baseexceptiongroup_task(ray_start_regular):
+@skip_if_python_less_than_3_11
+def test_baseexceptiongroup_task(ray_start_regular_shared):
     baseexceptiongroup = BaseExceptionGroup(  # noqa: F821
         "test baseexceptiongroup", [BaseException("abc")]
     )
@@ -25,7 +139,8 @@ def test_baseexceptiongroup_task(ray_start_regular):
         ray.get(task.remote())
 
 
-def test_baseexceptiongroup_actor(ray_start_regular):
+@skip_if_python_less_than_3_11
+def test_baseexceptiongroup_actor(ray_start_regular_shared):
     baseexceptiongroup = BaseExceptionGroup(  # noqa: F821
         "test baseexceptiongroup", [BaseException("abc")]
     )
@@ -40,7 +155,8 @@ def test_baseexceptiongroup_actor(ray_start_regular):
         ray.get(a.f.remote())
 
 
-def test_except_exceptiongroup(ray_start_regular):
+@skip_if_python_less_than_3_11
+def test_except_exceptiongroup(ray_start_regular_shared):
     exceptiongroup = ExceptionGroup(  # noqa: F821
         "test exceptiongroup", [ValueError(), TypeError()]
     )
@@ -74,7 +190,8 @@ def test_except_exceptiongroup(ray_start_regular):
         assert isinstance(ex.exceptions[1], TypeError)
 
 
-def test_except_star_exception(ray_start_regular):
+@skip_if_python_less_than_3_11
+def test_except_star_exception(ray_start_regular_shared):
     @ray.remote
     def task():
         raise ValueError
@@ -126,7 +243,8 @@ def test_except_star_exception(ray_start_regular):
     exec(python_code)
 
 
-def test_except_star_exceptiongroup(ray_start_regular):
+@skip_if_python_less_than_3_11
+def test_except_star_exceptiongroup(ray_start_regular_shared):
     exceptiongroup = ExceptionGroup(  # noqa: F821
         "test exceptiongroup", [ValueError(), TypeError()]
     )
