+++ python/ray/tests/test_failure.py
@@ -380,56 +380,6 @@ def test_exception_chain(ray_start_regular):
         assert isinstance(ex, RayTaskError)
 
 
-def test_baseexception_task(ray_start_regular):
-    class MyBaseException(BaseException):
-        pass
-
-    @ray.remote
-    def task():
-        raise MyBaseException("abc")
-
-    with pytest.raises(MyBaseException):
-        ray.get(task.remote())
-
-
-def test_baseexception_actor_task(ray_start_regular):
-    class MyBaseException(BaseException):
-        pass
-
-    @ray.remote
-    class Actor:
-        def f(self):
-            raise MyBaseException("abc")
-
-        async def async_f(self):
-            raise MyBaseException("abc")
-
-    a = Actor.remote()
-    with pytest.raises(MyBaseException):
-        ray.get(a.f.remote())
-
-    a = Actor.remote()
-    with pytest.raises(MyBaseException):
-        ray.get(a.async_f.remote())
-
-
-def test_baseexception_actor_creation(ray_start_regular):
-    class MyBaseException(BaseException):
-        pass
-
-    @ray.remote
-    class Actor:
-        def __init__(self):
-            raise MyBaseException("abc")
-
-    a = Actor.remote()
-    try:
-        ray.get(a.__ray_ready__.remote())
-        raise Exception("abc")
-    except ActorDiedError as e:
-        assert "MyBaseException" in str(e)
-
-
 @pytest.mark.skip("This test does not work yet.")
 @pytest.mark.parametrize("ray_start_object_store_memory", [10**6], indirect=True)
 def test_put_error1(ray_start_object_store_memory, error_pubsub):
